pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["agents", "logos", "shared"]

[project]
name = "hapax-officium"
version = "1.0.0"
description = "Management decision support system for engineering managers"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "jinja2>=3.1",
    "langfuse>=3.14.5",
    "matplotlib>=3.9",
    "ollama>=0.6.1",
    "pillow>=11.0.0",
    "playwright>=1.50.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.63.0",
    "pyyaml>=6.0",
    "qdrant-client>=1.17.0",
    "sse-starlette>=2.0.0",
    "uvicorn>=0.34.0",
    "watchdog>=4.0.0",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    "opentelemetry-exporter-otlp-proto-http>=1.39.0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "hapax-sdlc[demo] @ git+https://github.com/ryanklee/hapax-constitution.git@cbdf204756fc6f0583d3056669e2c8643ca15d21",
]

[project.scripts]
logos = "logos.__main__:main"
logos-api = "logos.api.__main__:main"

[dependency-groups]
dev = [
    "pyright>=1.1.390",
    "pytest>=9.0.2",
    "pytest-cov>=6.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
extend = "ruff.toml"
